requirements.txt
newsapi-python
python-dotenv
sentence-transformers
pandas
transformers
torch